gumm-local/server/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
pydantic>=2.5.0
email-validator>=2.0.0
